# [build] Run class-parse and api-xml-adjuster in parallel