--- a/build-tools/Xamarin.Android.Tools.BootstrapTasks/Xamarin.Android.Tools.BootstrapTasks/RunParallelCmds.cs
+++ b/build-tools/Xamarin.Android.Tools.BootstrapTasks/Xamarin.Android.Tools.BootstrapTasks/RunParallelCmds.cs
@@ -24,6 +24,8 @@ namespace Xamarin.Android.Tools.BootstrapTasks
 		{
 			bool success = true;
 
+			LogMessage ($"RunParallelCmds max parallelism: {Environment.ProcessorCount}");
+
 			Tasks.Parallel.ForEach (Commands, new Tasks.ParallelOptions () { MaxDegreeOfParallelism = Environment.ProcessorCount },
 				cmd => {
 					var command = cmd.GetMetadata ("Command");