Improve the build speed by running the class-parse and api-xml-adjuster tools in parallel.

Measured times:

Original:

    Project Performance Summary:
        60265 ms  C:\Users\rodo\git\xamarin-android\build-tools\api-xml-adjuster\api-xml-adjuster.csproj   1 calls
                  60265 ms  _AdjustApiXml                              1 calls
    
    Target Performance Summary:
           20 ms  _DefineApiFiles                            1 calls
        26117 ms  _AdjustApiXml                             16 calls
        34116 ms  _ClassParse                               16 calls
    
    Task Performance Summary:
            3 ms  MakeDir                                   16 calls
            5 ms  CreateItem                                16 calls
        60221 ms  Exec                                      32 calls

New:

    Project Performance Summary:
         5673 ms  C:\Users\rodo\git\xamarin-android\build-tools\api-xml-adjuster\api-xml-adjuster.csproj   1 calls
                   5673 ms  _AdjustApiXml                              1 calls
    
    Target Performance Summary:
            3 ms  _AdjustApiXmlPrepare                       1 calls
            4 ms  _ClassParsePrepare                         1 calls
           14 ms  _DefineApiFiles                            1 calls
         2437 ms  _AdjustApiXml                              1 calls
         3200 ms  _ClassParse                                1 calls
    
    Task Performance Summary:
            1 ms  MakeDir                                    1 calls
            7 ms  CreateItem                                48 calls
         5633 ms  RunParallelCmds                            2 calls

CI builds:

mac before: `make jenkins ... 31m 41s` (https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3538996) 
mac after: `make jenkins ... 29m 51s` (https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3539034)

windows before: `msbuild Xamarin.Android /t:Build ... 11m 18s` (https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3538996)
windows after: `msbuild Xamarin.Android /t:Build ... 11m 59s` (https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3539034 slower, hopefully it is build bot variance?)