@jonpryor 2020-03-10
It failed to build :-(

https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3537045&view=logs&j=96fd57f5-f69e-53c7-3d47-f67e6cf9b93e&t=65256bb7-a34c-5353-bc4d-c02ee25dc133

```
  /Users/builder/azdo/_work/3/s/xamarin-android/src/Xamarin.Android.NUnitLite/Xamarin.Android.NUnitLite.csproj : error MSB4057: The target "Build" does not exist in the project.

```