@radekdoulik 2020-03-10
> It failed to build :-(

The mac build looks quite broken :-) Looks like that's what one get for trying to switch to windows...